Split settings-page container into multiple components
