+++ src/app/containers/index.ts
@@ -2,7 +2,6 @@ export * from './check-in/check-in.component';
 export * from './getting-started/getting-started.component';
 export * from './login/login.component';
 export * from './organization-dashboard/organization-dashboard.component';
-export * from '../settings/containers/settings-page/settings-page.component';
 export * from './join-page/join-page.component';
 export * from './public-organization-dashboard/public-organization-dashboard.component';
 export * from './find-organization/find-organization.component';
