--- src/app/containers/organization-settings/organization-settings.component.html
@@ -1,3 +1,0 @@
-<p>
-  organization-settings works!
-</p>
