--- src/app/containers/organization-settings/organization-settings.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-organization-settings',
-  templateUrl: './organization-settings.component.html',
-  styleUrls: ['./organization-settings.component.scss']
-})
-export class OrganizationSettingsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
