--- src/app/settings/components/settings-layout/settings-layout.component.html
@@ -1,3 +1,0 @@
-<p>
-  settings-layout works!
-</p>
