--- src/app/settings/components/settings-layout/settings-layout.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-settings-layout',
-  templateUrl: './settings-layout.component.html',
-  styleUrls: ['./settings-layout.component.scss']
-})
-export class SettingsLayoutComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
