+++ src/app/settings/containers/roles/roles.component.html
@@ -1,3 +1,6 @@
-<p>
-  roles works!
-</p>
+<app-data-table
+  [data$]="users$"
+  [columnOptions]="userTableOptions"
+  (deleteItem)="onDeleteUser($event)"
+  (updateItem)="onUpdateUser($event)">
+</app-data-table>
