+++ src/app/settings/containers/settings-page/settings-page.component.html
@@ -1,63 +1,23 @@
 <!--User form-->
-<div *ngIf="sidenavSelection === 'user'">
-  <app-user-form
-    [title]="userFormTitle"
-    [initialUser]="currentUser"
-    [passwordRequired]="false"
-    (validUser)="onValidUser($event, currentUser)">
-  </app-user-form>
-  <div class="container-submit">
-    <button
-      mat-raised-button
-      color="primary"
-      (click)="onSubmitUser(userChanges)"
-      [disabled]="!userChanges">
-      Submit
-    </button>
-  </div>
-</div>
-<!--Organization form-->
-<div *ngIf="sidenavSelection === 'organization'">
-  <app-organization-form
-    [title]="organizationFormTitle"
-    [initialOrganization]="currentOrganization"
-    (validOrganization)="onValidOrganization($event, currentOrganization)">
-  </app-organization-form>
-  <div class="container-submit">
-    <button
-      mat-raised-button
-      color="primary"
-      (click)="onSubmitOrganization(organizationChanges)"
-      [disabled]="!organizationChanges">
-      Submit
-    </button>
-  </div>
-</div>
-<!--Volunteer data table-->
-<div *ngIf="sidenavSelection === 'volunteers'">
-  <app-data-table
-    [data$]="volunteers$"
-    [columnOptions]="volunteerTableOptions"
-    [showDelete]="true"
-    (deleteItem)="onDeleteVolunteer($event)">
-  </app-data-table>
-</div>
-<!--Reports form-->
-<div *ngIf="sidenavSelection == 'reports'">
-  <app-reports-form
-    (validReport)="onValidReport($event)">
-  </app-reports-form>
-  <div class="container-submit">
-    <button mat-raised-button color="primary" [disabled]="!validReport" (click)="onSubmitReport()">Export Report
-    </button>
-  </div>
-</div>
-<!--Roles data table-->
-<div *ngIf="sidenavSelection == 'roles'">
-  <app-data-table
-    [data$]="users$"
-    [columnOptions]="userTableOptions"
-    (deleteItem)="onDeleteUser($event)"
-    (updateItem)="onUpdateUser($event)">
-  </app-data-table>
+<div [ngSwitch]="sidenavSelection">
+  <app-user-settings
+    *ngSwitchCase="'USER_SETTINGS'"
+  >
+  </app-user-settings>
+  <app-organization-settings
+    *ngSwitchCase="'ORGANIZATION_SETTINGS'"
+  >
+  </app-organization-settings>
+  <app-volunteer-settings
+    *ngSwitchCase="'VOLUNTEER_SETTINGS'"
+  >
+  </app-volunteer-settings>
+  <app-roles
+    *ngSwitchCase="'ROLES'"
+  >
+  </app-roles>
+  <app-reports
+    *ngSwitchCase="'REPORTS'"
+  >
+  </app-reports>
 </div>
