+++ src/app/settings/containers/user-settings/user-settings.component.html
@@ -1,3 +1,15 @@
-<p>
-  user-settings works!
-</p>
+<app-user-form
+  [title]="userFormTitle"
+  [initialUser]="currentUser"
+  [passwordRequired]="false"
+  (validUser)="onValidUser($event, currentUser)">
+</app-user-form>
+<div class="container-submit">
+  <button
+    mat-raised-button
+    color="primary"
+    (click)="onSubmitUser(userChanges)"
+    [disabled]="!userChanges">
+    Submit
+  </button>
+</div>
