+++ src/app/settings/containers/volunteer-settings/volunteer-settings.component.html
@@ -1,3 +1,6 @@
-<p>
-  volunteer-settings works!
-</p>
+  <app-data-table
+    [data$]="volunteers$"
+    [columnOptions]="volunteerTableOptions"
+    [showDelete]="true"
+    (deleteItem)="onDeleteVolunteer($event)">
+  </app-data-table>
